% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_factanal.R
\name{ex_factanal}
\alias{ex_factanal}
\title{Extract SS loadings, Proportion Var and Cumulative Var from factanal output}
\usage{
ex_factanal(factanal_object)
}
\arguments{
\item{factanal_object}{Output from \emph{factanal()} call}
}
\description{
Extract SS loadings, Proportion Var and Cumulative Var from factanal output
}
