% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load2.R
\name{load2}
\alias{load2}
\title{Load file and set its variable's name in one go}
\usage{
load2(file)
}
\arguments{
\item{file}{Path to the input file}
}
\description{
Load file and set its variable's name in one go
}
