% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerPar.R
\name{registerPar}
\alias{registerPar}
\title{Register cores for parallel processing on Unix & Windows}
\usage{
registerPar()
}
\description{
Register cores for parallel processing on Unix & Windows
}
\note{
just a wrap for unix and windows systems. DoParallel should theoretically work on both, though. Will leave 1 core for os if not run on r-workbench.
}
