% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_impute_constant.R
\name{step_impute_constant}
\alias{step_impute_constant}
\title{Impute a constant value to correct for missingness}
\usage{
step_impute_constant(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  constant = 0,
  skip = FALSE,
  id = rand_id("impute_constant")
)
}
\arguments{
\item{role}{}

\item{constant}{A numeric value or a possibly named vector with same length
as the result from the selector functions, specifying which constant should
be imputed for each selector}
}
\description{
This method can be used for constant imputation.
}
\note{
taken from https://github.com/tidymodels/recipes/issues/473, edited by
me to make it work.
}
\author{
Bijaelo
}
