% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_outliers_iqr__to_limits.R
\name{step_outliers_iqr_to_limits}
\alias{step_outliers_iqr_to_limits}
\title{Identify values outside the IQR and replace them with IQR limits}
\usage{
step_outliers_iqr_to_limits(
  recipe,
  ...,
  role = NA,
  skip = TRUE,
  trained = FALSE,
  columns = NULL,
  id = recipes::rand_id("outliers_iqr_to_limits")
)
}
\description{
This method can be used in order to identify outliers by using
the interquartile range rule and replace them with min/max values of its
range.
}
\note{
I think it's quite cool.
}
\author{
steffanossa
}
