% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_unwanted2NA.R
\name{step_unwanted2NA}
\alias{step_unwanted2NA}
\title{Turn unwanted values into NAs}
\usage{
step_unwanted2NA(
  recipe,
  ...,
  role = NA,
  skip = TRUE,
  values = c(0, 1),
  wanted = TRUE,
  trained = FALSE,
  columns = NULL,
  id = recipes::rand_id("unwanted2NA")
)
}
\arguments{
\item{values}{Single value or vector containing values wanted in columns
set.}

\item{wanted}{Boolean. Turns everything not in \emph{values} into NAs if \emph{TRUE}.
Turn everything in \emph{values} into NAs if \emph{FALSE}.}
}
\description{
This method can be used in order to cleanse columns from values
that are unwanted.
}
\note{
I think it's quite cool.

IMPORTANT: use prep(strings_as_factors = FALSE) when dealing with chr
columns! Otherwise you will end up with chr columns turned to int!
}
\author{
steffanossa
}
